move ResultSetMapping to a better package and add javadoc
