--- hibernate-reactive-core/src/main/java/org/hibernate/reactive/ResultSetMapping.java
@@ -1,11 +1,0 @@
-/* Hibernate, Relational Persistence for Idiomatic Java
- *
- * SPDX-License-Identifier: LGPL-2.1-or-later
- * Copyright: Red Hat Inc. and Hibernate Authors
- */
-package org.hibernate.reactive;
-
-public interface ResultSetMapping<T> {
-	String getName();
-	Class<T> getResultType();
-}
